--- a/Cargo.toml
+++ b/Cargo.toml
@@ -89,3 +89,20 @@ serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
 time = "0.1"
+
+[patch.crates-io]
+tokio = { path = "." }
+tokio-async-await  = { path = "./tokio-async-await" }
+tokio-codec = { path = "./tokio-codec" }
+tokio-current-thread = { path = "./tokio-current-thread" }
+tokio-executor = { path = "./tokio-executor" }
+tokio-fs = { path = "./tokio-fs" }
+tokio-io = { path = "./tokio-io" }
+tokio-reactor = { path = "./tokio-reactor" }
+tokio-signal = { path = "./tokio-signal" }
+tokio-tcp = { path = "./tokio-tcp" }
+tokio-threadpool = { path = "./tokio-threadpool" }
+tokio-timer = { path = "./tokio-timer" }
+tokio-tls = { path = "./tokio-tls" }
+tokio-udp = { path = "./tokio-udp" }
+tokio-uds = { path = "./tokio-uds" }