@jonhoo 2018-10-03
Interesting... `tokio-io-pool` uses a slightly different heuristic here: anything that gets spawned *onto* the pool (e.g. with `Runtime::spawn`, `Runtime::spawn_all`, or through a `Handle`) is sent to a random worker, whereas anything that is spawned *on* the pool (through `tokio::spawn`). The idea behind this was that when you add a *new* thing onto the pool, it's likely that that that's the result of something like an `accept`, and then it should definitely be distributed evenly (as you observe in this PR). Once something is *on* the pool though, we want to keep `spawn` fast, so we may as well spawn it locally. Specifically, the observation is that I/O tasks don't generally spawn lots of *other* I/O tasks.

This might be trickier in `tokio` proper since the listener will *also* be running on the pool, so we don't have a good way of distinguishing the two...