@stjepang 2018-10-03
> This might be trickier in tokio proper since the listener will also be running on the pool, so we don't have a good way of distinguishing the two...

Yep, that's why I'm leaning towards always spawning a new task onto a random worker - just in case.

> Once something is on the pool though, we want to keep spawn fast, so we may as well spawn it locally.

I'm not as experienced with the real-world use of Tokio, but I wonder: Is the cost of spawning something we should be particularly worried about?

My gut feeling is that tasks are relatively rarely spawned and that the cost of spawning tends to be negligible in the grand scheme of things. On the other hand, I'd wager tasks are yielded much more often than they're spawned so we should strive to optimize task switching as much as possible.