@jonhoo 2018-10-03
Yeah, you could be right that spawning is relatively rare on the critical path, so this may just be a non-issue :)