@carllerche 2018-10-03
I'm using this patch to bench the [hyper hello example](https://github.com/hyperium/hyper/blob/master/examples/hello.rs):

### With this patch:

```
$ wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   368.86us   69.77us   4.96ms   90.77%
    Req/Sec    75.06k     6.64k   79.88k    88.00%
  746917 requests in 10.00s, 62.68MB read
Requests/sec:  74679.24
Transfer/sec:      6.27MB
```

**Requests/sec:  74679.24**

### Against `current_thread::Runtime`

```
wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   603.60us   84.34us   5.56ms   91.32%
    Req/Sec    81.17k     2.25k   82.73k    98.00%
  807192 requests in 10.00s, 67.74MB read
Requests/sec:  80708.63
Transfer/sec:      6.77MB
```

**Requests/sec:  80708.63**

So, it seems like there is still some source of additional overhead there.