@carllerche 2018-10-03
Hyper is doing basic spawning [here](https://github.com/hyperium/hyper/blob/168c7d2155952ba09f781c331fd67593b820af20/src/common/exec.rs#L25), so in theory it should be distributed across the runtime.

It could be that multi threading provides no added benefit in this benchmark, but I am not 100% sure what is going on.