@stjepang 2018-10-03
Here's the same hyper `hello` benchmark on my machine.

Baseline `tokio-io-pool`:

```
$ wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   362.54us  603.80us  14.48ms   97.47%
    Req/Sec    91.21k     3.24k   96.54k    72.00%
  907058 requests in 10.00s, 76.12MB read
Requests/sec:  90675.35
Transfer/sec:      7.61MB
```

`tokio-threadpool` before PR #660:

```
$ wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   452.42us  743.86us  26.01ms   96.34%
    Req/Sec    77.83k     3.30k   85.16k    83.00%
  774089 requests in 10.00s, 64.96MB read
Requests/sec:  77374.46
Transfer/sec:      6.49MB
```

PR #660:

```
$ wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   418.65us    0.90ms  18.85ms   97.56%
    Req/Sec    86.94k     2.85k   92.42k    78.00%
  865452 requests in 10.01s, 72.63MB read
Requests/sec:  86487.88
Transfer/sec:      7.26MB
```

This PR:

```
$ wrk -t1 -d10 -c50 http://127.0.0.1:3000/
Running 10s test @ http://127.0.0.1:3000/
  1 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   420.79us  812.86us  18.17ms   97.32%
    Req/Sec    87.12k     3.24k   95.14k    71.00%
  867375 requests in 10.01s, 72.79MB read
Requests/sec:  86687.78
Transfer/sec:      7.28MB
```