@stjepang 2018-10-03
My machine:

* OS: GNU/Linux, kernel v4.17.19
* Arch: x86-64
* CPU: Intel(R) Core(TM) i7-5600U CPU @ 2.60GHz
* Cores: 2 physical / 4 logical